Start debug of destructors segfault
Works for step1 on GPU and CPU but on GPU it fails to destroy and
freeing stops working does not seem to be connected to a segment to be
freed just after a certain point it stops working it seems like as I can
comment out the free that fails and then the next free will fail.

So it might be that the context was destroyed. The PopSift class which
initially created the context is freed first so not sure if that is the
cause or if it is something else