Working Bindless for input (upscale)
Tends to have one less extrema than cuda and USM (sometimes the same)
but I think that is still considered correct USM was not always
identical to cuda either. USM does now faild to run the complete 957
images in the AI vs Human dataset that it was able to run before so
seems like some bug have been introduced. Bindless verions is able to
run that complete dataset