Attempt to do matching with joint matrix extension
Not currently working due to loading from local memory after transpose.
Might work to load directly from meory as col-major but don't think that
handles the transpose. And if it does not shared memory is the only
viable option too still make it efficient(probably)

Would in most cases be slwoer on strong GPU's anyways due to not being
enought extremas in normal case to saturate the gpu